+++ gradle/libs.versions.toml
@@ -76,7 +76,6 @@ jsoup = { module = "org.jsoup:jsoup", version.ref = "jsoup" }
 junit-bom = { group = "org.junit", name = "junit-bom", version.ref = "junit" }
 junit-jupiter = { group = "org.junit.jupiter", name = "junit-jupiter", version.ref = "junit" }
 junit-platform-launcher = { group = "org.junit.platform", name = "junit-platform-launcher", version.ref = "junitPlatformLauncher" }
-junit-platform-suite = { group = "org.junit.platform", name = "junit-platform-suite", version.ref = "junitPlatformLauncher" }
 mockk = { group = "io.mockk", name = "mockk", version.ref = "mockk" }
 androidx-junit-ktx = { group = "androidx.test.ext", name = "junit-ktx", version.ref = "junitKtx" }
 navigation-testing = { group = "androidx.navigation", name = "navigation-testing", version.ref = "nav_version" }
